minor version bump
Version checker; session id for avoiding accidental shutdown/reboot;

+++ install.sh
@@ -9,20 +9,75 @@ fi
 
 
 if [ "$1" == "update" ]; then
-    echo "Backing up old Parameters and WiFi"
-    cp /home/pi/bcMeterConf.py  /home/pi/bcMeterConf.orig
-    cp /home/pi/bcMeter_wifi.json  /home/pi/bcMeter_wifi.json.orig
-    echo "Updating from github"
-    rm -rf /home/pi/bcmeter/ /home/pi/interface/
-    git clone https://github.com/bcmeter/bcmeter.git /home/pi/bcmeter
-    mv /home/pi/bcmeter/* /home/pi/
-    rm -rf /home/pi/gerbers/ /home/pi/stl/
+    version=''
+    localfile=bcMeter.py
 
-    echo "Restoring Parameters"
-    mv /home/pi/bcMeterConf.orig /home/pi/bcMeterConf.py
-    mv /home/pi/bcMeter_wifi.json.orig /home/pi/bcMeter_wifi.json  
+    for((i=1; i<=50; i++))
+    do
+      line=$(head -$i $localfile | tail -1)
+      if [[ $line == bcMeter_version* ]]
+      then
+        version=$(echo $line | cut -d '"' -f2)
+        break
+      else
+         version='0.9.20' #every version before was not set as variable so it has to be older. force update then.
+      fi
+    done
 
+    version_parts=($(echo $version | tr '.' ' '))
 
+    major_version=${version_parts[0]}
+    minor_version=${version_parts[1]}
+    patch_version=${version_parts[2]}
+
+    echo "installed version: " $major_version $minor_version $patch_version
+    wget -q  -P /tmp/ https://raw.githubusercontent.com/dahljo/bcmeter/main/bcMeter.py
+
+    localfile=/tmp/bcMeter.py
+    echo "checking" $localfile
+
+    for((i=1; i<=50; i++))
+    do
+      line=$(head -$i $localfile | tail -1)
+      if [[ $line == bcMeter_version* ]]
+      then
+        version=$(echo $line | cut -d '"' -f2)
+        break
+        else
+         version='0.9.19' #if not found more recent online, dont update
+      fi
+
+    done
+
+    version_parts=($(echo $version | tr '.' ' '))
+
+    git_major_version=${version_parts[0]}
+    git_minor_version=${version_parts[1]}
+    git_patch_version=${version_parts[2]}
+
+    echo "available version: " $git_major_version $git_minor_version $git_patch_version
+    rm /tmp/bcMeter*
+    if [ "$major_version" -lt "$git_major_version" ] || [ "$minor_version" -lt "$git_minor_version" ] || [ "$patch_version" -lt "$git_patch_version" ]
+    then
+        echo "running updating"
+        systemctl stop bcMeter_ap_control_loop
+        systemctl stop bcMeter
+        rm /home/pi/ap_control_loop.log
+        echo "Backing up old Parameters and WiFi"
+        cp /home/pi/bcMeterConf.py  /home/pi/bcMeterConf.orig
+        cp /home/pi/bcMeter_wifi.json  /home/pi/bcMeter_wifi.json.orig
+        echo "Updating from github"
+        rm -rf /home/pi/bcmeter/ /home/pi/interface/
+        git clone https://github.com/bcmeter/bcmeter.git /home/pi/bcmeter
+        mv /home/pi/bcmeter/* /home/pi/
+        rm -rf /home/pi/gerbers/ /home/pi/stl/
+        echo "Restoring Parameters"
+        mv /home/pi/bcMeterConf.orig /home/pi/bcMeterConf.py
+        mv /home/pi/bcMeter_wifi.json.orig /home/pi/bcMeter_wifi.json  
+        systemctl start bcMeter_ap_control_loop
+        systemctl start bcMeter
+     else echo "most recent version installed"
+    fi
 fi
 
 
@@ -37,7 +92,7 @@ apt update && apt upgrade -y && apt autoremove -y;
         fi
 
     echo "Installing software packages needed to run bcMeter. This will take a while and is dependent on your internet connection, the amount of updates and the speed of your pi."
-    apt install -y i2c-tools zram-tools python3-pip python3-smbus python3-dev python3-rpi.gpio python3-numpy nginx php php-fpm php-pear php-common php-cli php-gd screen git openssl && pip3 install gpiozero adafruit-blinka tabulate && systemctl enable zramswap.service  
+    apt install -y i2c-tools zram-tools python3-pip python3-smbus python3-dev python3-rpi.gpio python3-numpy nginx php php-fpm php-pear php-common php-cli php-gd screen git openssl && pip3 install gpiozero tabulate && systemctl enable zramswap.service  
 
 
     GITCLONE=0
@@ -121,15 +176,10 @@ fi
 touch $BCMINSTALLED 
 if [ "$1" != "update" ]; then
     read -p "Basically the bcMeter is now set up. It is recommended to install the WiFi Accesspoint. It will create an own WiFi if no known WiFi is available. Continue? " yn
-        case $yn in
-            [Yy]* ) bash accesspoint-install.sh; break;;
-            [Nn]* ) exit;;
-            * ) echo "Please answer yes (y) or no (n).";;
-        esac
-    fi
+    case $yn in
+        [Yy]* ) bash accesspoint-install.sh; break;;
+        [Nn]* ) exit;;
+        * ) echo "Please answer yes (y) or no (n).";;
+    esac
 fi
-
-
-
-
